requirements.txt
# Core
python-dotenv>=1.0.1
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.7.1

# NLP & Preprocessing
nltk>=3.8.1
spacy>=3.7.4
deepmultilingualpunctuation>=1.0.3
symspellpy>=6.7.7
transformers>=4.40.1
sentencepiece>=0.2.0
torch>=2.2.2
soundfile>=0.12.1

# Transcription
SpeechRecognition>=3.10.1
pydub>=0.25.1

# Web UI
streamlit>=1.34.0
requests>=2.31.0

# Logging & Utilities
loguru>=0.7.2
python-multipart>=0.0.9

# For uploading & audio I/O
aiofiles>=23.2.1

# If using Google Speech-to-Text
google-cloud-speech>=2.26.0
